integrated entity framework to postresql with entities configurations.

+++ Yenilen.Application/Interfaces/IReviewRepository.cs
@@ -5,6 +5,6 @@ namespace Yenilen.Application.Interfaces;
 public interface IReviewRepository:IGenericRepository<Review>
 {
     Task<int> GetReviewCountByStoreId(int id);
-    Task<float> GetStoreRatingByStoreId(int id);
+    Task<decimal> GetStoreRatingByStoreId(int id);
 
 }
